@@ -1,77 +1,39 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  darkMode: ["class"],
+export default {
   content: [
-    './pages/**/*.{js,jsx}',
-    './components/**/*.{js,jsx}',
-    './app/**/*.{js,jsx}',
-    './src/**/*.{js,jsx}',
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
-  prefix: "",
   theme: {
-    container: {
-      center: true,
-      padding: "2rem",
-      screens: {
-        "2xl": "1400px",
-      },
-    },
     extend: {
       colors: {
-        border: "hsl(var(--border))",
-        input: "hsl(var(--input))",
-        ring: "hsl(var(--ring))",
-        background: "hsl(var(--background))",
-        foreground: "hsl(var(--foreground))",
-        primary: {
-          DEFAULT: "hsl(var(--primary))",
-          foreground: "hsl(var(--primary-foreground))",
-        },
-        secondary: {
-          DEFAULT: "hsl(var(--secondary))",
-          foreground: "hsl(var(--secondary-foreground))",
-        },
-        destructive: {
-          DEFAULT: "hsl(var(--destructive))",
-          foreground: "hsl(var(--destructive-foreground))",
-        },
-        muted: {
-          DEFAULT: "hsl(var(--muted))",
-          foreground: "hsl(var(--muted-foreground))",
-        },
-        accent: {
-          DEFAULT: "hsl(var(--accent))",
-          foreground: "hsl(var(--accent-foreground))",
-        },
-        popover: {
-          DEFAULT: "hsl(var(--popover))",
-          foreground: "hsl(var(--popover-foreground))",
-        },
-        card: {
-          DEFAULT: "hsl(var(--card))",
-          foreground: "hsl(var(--card-foreground))",
-        },
+        'background': 'hsl(var(--background) / <alpha-value> )',
+        'foreground': 'hsl(var(--foreground) / <alpha-value> )',
+        'card': 'hsl(var(--card) / <alpha-value> )',
+        'card-foreground': 'hsl(var(--card-foreground) / <alpha-value> )',
+        'popover': 'hsl(var(--popover) / <alpha-value> )',
+        'popover-foreground': 'hsl(var(--popover-foreground) / <alpha-value> )',
+        'primary': 'hsl(var(--primary) / <alpha-value> )',
+        'primary-foreground': 'hsl(var(--primary-foreground) / <alpha-value> )',
+        'secondary': 'hsl(var(--secondary) / <alpha-value> )',
+        'secondary-foreground': 'hsl(var(--secondary-foreground) / <alpha-value> )',
+        'muted': 'hsl(var(--muted) / <alpha-value> )',
+        'muted-foreground': 'hsl(var(--muted-foreground) / <alpha-value> )',
+        'accent': 'hsl(var(--accent) / <alpha-value> )',
+        'accent-foreground': 'hsl(var(--accent-foreground) / <alpha-value> )',
+        'destructive': 'hsl(var(--destructive) / <alpha-value> )',
+        'destructive-foreground': 'hsl(var(--destructive-foreground) / <alpha-value> )',
+        'border': 'hsl(var(--border) / <alpha-value> )',
+        'input': 'hsl(var(--input) / <alpha-value> )',
+        'ring': 'hsl(var(--ring) / <alpha-value> )',
       },
-      borderRadius: {
-        lg: "var(--radius)",
-        md: "calc(var(--radius) - 2px)",
-        sm: "calc(var(--radius) - 4px)",
+      backgroundImage: {
+        'logo-image': "url('src/assets/images/logo.webp')",
       },
-      keyframes: {
-        "accordion-down": {
-          from: { height: "0" },
-          to: { height: "var(--radix-accordion-content-height)" },
-        },
-        "accordion-up": {
-          from: { height: "var(--radix-accordion-content-height)" },
-          to: { height: "0" },
-        },
-      },
-      animation: {
-        "accordion-down": "accordion-down 0.2s ease-out",
-        "accordion-up": "accordion-up 0.2s ease-out",
-      },
+      fontFamily: {
+        Geist: ['Geist', 'ui-sans-serif', 'system-ui'],
+      }
     },
   },
-  plugins: [require("tailwindcss-animate")],
+  plugins: [],
 }
